+++ src/adapter/http/src/data/tail_handler.rs
@@ -48,8 +48,10 @@ pub async fn dataset_tail_handler(
         .await
         .map_err(|e| match e {
             QueryError::DatasetNotFound(e) => ApiError::not_found(e),
+            QueryError::DatasetBlockNotFound(_) | QueryError::DataFusionError(_) => {
+                e.int_err().api_err()
+            }
             QueryError::DatasetSchemaNotAvailable(e) => ApiError::no_content(e),
-            QueryError::DataFusionError(e) => e.int_err().api_err(),
             QueryError::Access(e) => e.api_err(),
             QueryError::Internal(e) => e.api_err(),
         })?;