+++ src/adapter/http/tests/tests/test_data_query.rs
@@ -13,6 +13,7 @@ use chrono::{TimeZone, Utc};
 use datafusion::arrow::array::{RecordBatch, StringArray, UInt64Array};
 use datafusion::arrow::datatypes::*;
 use datafusion::prelude::*;
+use ed25519_dalek::Signer;
 use kamu::domain::*;
 use kamu::*;
 use kamu_adapter_http::data::query_types::IdentityConfig;
@@ -32,6 +33,7 @@ struct Harness {
     root_url: url::Url,
     dataset_handle: DatasetHandle,
     dataset_url: url::Url,
+    private_key: PrivateKey,
 }
 
 impl Harness {
@@ -39,11 +41,11 @@ impl Harness {
         // TODO: Need access to these from harness level
         let run_info_dir = tempfile::tempdir().unwrap();
 
+        let private_key: PrivateKey =
+            ed25519_dalek::SigningKey::from_bytes(&[123; ed25519_dalek::SECRET_KEY_LENGTH]).into();
+
         let identity_config = IdentityConfig {
-            private_key: ed25519_dalek::SigningKey::from_bytes(
-                &[123; ed25519_dalek::SECRET_KEY_LENGTH],
-            )
-            .into(),
+            private_key: private_key.clone(),
         };
 
         let catalog = dill::CatalogBuilder::new()
@@ -166,6 +168,7 @@ impl Harness {
             root_url,
             dataset_handle: create_result.dataset_handle,
             dataset_url,
+            private_key,
         }
     }
 }
@@ -367,7 +370,7 @@ async fn test_data_query_handler_v2() {
                         {"city": "B", "offset": 1, "population": 200},
                         {"city": "A", "offset": 0, "population": 100},
                     ],
-                    "dataFormat": "JsonAos",
+                    "dataFormat": "JsonAoS",
                 }
             })
         );
@@ -395,7 +398,7 @@ async fn test_data_query_handler_v2() {
                     "include": ["Input", "Schema"],
                     "query": query,
                     "queryDialect": "SqlDataFusion",
-                    "dataFormat": "JsonAos",
+                    "dataFormat": "JsonAoS",
                     "schemaFormat": "ArrowJson",
                     "skip": 0,
                     "limit": 100,
@@ -410,7 +413,7 @@ async fn test_data_query_handler_v2() {
                         {"city": "B", "offset": 1, "population": 200},
                         {"city": "A", "offset": 0, "population": 100},
                     ],
-                    "dataFormat": "JsonAos",
+                    "dataFormat": "JsonAoS",
                     "schema": ignore_schema,
                     "schemaFormat": "ArrowJson",
                 }
@@ -441,7 +444,7 @@ async fn test_data_query_handler_v2() {
                     "include": ["Input", "Proof", "Schema"],
                     "query": query,
                     "queryDialect": "SqlDataFusion",
-                    "dataFormat": "JsonAos",
+                    "dataFormat": "JsonAoS",
                     "schemaFormat": "ArrowJson",
                     "skip": 0,
                     "limit": 100,
@@ -456,20 +459,20 @@ async fn test_data_query_handler_v2() {
                         {"city": "B", "offset": 1, "population": 200},
                         {"city": "A", "offset": 0, "population": 100},
                     ],
-                    "dataFormat": "JsonAos",
+                    "dataFormat": "JsonAoS",
                     "schema": ignore_schema,
                     "schemaFormat": "ArrowJson",
                 },
                 "subQueries": [],
                 "commitment": {
-                    "inputHash": "f162001ff67ca8970bcb4f4f8b25e79b3c6db3fcd2ac0501d131e446591fd0475a2af",
-                    "outputHash": "f1620fa841fae69710c888fdf82d8fd63948469c0fd1e2a37c16e2067127e2eec1ea8",
+                    "inputHash": "f1620a34bf5f8e0dbbe4ae6aefb60a3709d4eaaebd99173c90a0582c1b0011305a752",
+                    "outputHash": "f1620ba3be8f1b2ff3d37efc7051b43a1a78f635bb657874326ac394f99e486c6569d",
                     "subQueriesHash": "f1620ca4510738395af1429224dd785675309c344b2b549632e20275c69b15ed1d210",
                 },
                 "proof": {
                     "type": "Ed25519Signature2020",
                     "verificationMethod": "did:key:z6Mko2nqhQ9wYSTS5Giab2j1aHzGnxHimqwmFeEVY8aNsVnN",
-                    "proofValue": "u-k8Rd9dB5ERqbTU9ymUvpySTQEh8HMPAqcEBrtZviNOBFoe-FXZtJUGcvwud39dxC659bkVz4iYHhDYUexmiCQ",
+                    "proofValue": "u2FwMK6jHGML0EBQ8X3Q7gfwUGLzZBe6jJeYRug6jDoZM8lvEqM_fUk9itia0htm7vRZ8MD_fezG2sgBCc8e-Bw",
                 }
             })
         );
@@ -512,6 +515,333 @@ async fn test_data_query_handler_v2() {
         let commitment = canonical_json::to_string(&response["commitment"]).unwrap();
 
         did.verify(commitment.as_bytes(), &signature).unwrap();
+
+        // Error: Dataset does not exist
+        let res = cl
+            .post(&format!("{}query", harness.root_url))
+            .json(&json!({
+                "query": query,
+                "datasets": [{
+                    "id": DatasetID::new_seeded_ed25519(b"does-not-exist"),
+                    "alias": harness.dataset_handle.alias,
+                }],
+            }))
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::NOT_FOUND);
+
+        // Error: Block does not exist
+        let res = cl
+            .post(&format!("{}query", harness.root_url))
+            .json(&json!({
+                "query": query,
+                "datasets": [{
+                    "id": harness.dataset_handle.id,
+                    "alias": harness.dataset_handle.alias,
+                    "blockHash": Multihash::from_digest_sha3_256(b"does-not-exist"),
+                }],
+            }))
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::NOT_FOUND);
+    };
+
+    await_client_server_flow!(harness.server_harness.api_server_run(), client);
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(engine, datafusion)]
+#[test_log::test(tokio::test)]
+async fn test_data_verify_handler() {
+    let harness = Harness::new().await;
+
+    let client = async move {
+        let cl = reqwest::Client::new();
+
+        let head = cl
+            .get(format!("{}/refs/head", harness.dataset_url))
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap()
+            .text()
+            .await
+            .unwrap();
+
+        let query = format!(
+            "select offset, city, population from \"{}\" order by offset desc",
+            harness.dataset_handle.alias
+        );
+
+        // Get response with proof
+        let res = cl
+            .post(&format!("{}query", harness.root_url))
+            .json(&json!({
+                "query": query,
+                "include": ["proof"],
+            }))
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap();
+
+        let response = res.json::<serde_json::Value>().await.unwrap();
+
+        pretty_assertions::assert_eq!(
+            response,
+            json!({
+                "input": {
+                    "include": ["Input", "Proof"],
+                    "query": query,
+                    "queryDialect": "SqlDataFusion",
+                    "dataFormat": "JsonAoS",
+                    "skip": 0,
+                    "limit": 100,
+                    "datasets": [{
+                        "alias": "kamu-server/population",
+                        "blockHash": head,
+                        "id": harness.dataset_handle.id.as_did_str().to_string(),
+                    }],
+                },
+                "output": {
+                    "data": [
+                        {"city": "B", "offset": 1, "population": 200},
+                        {"city": "A", "offset": 0, "population": 100},
+                    ],
+                    "dataFormat": "JsonAoS",
+                },
+                "subQueries": [],
+                "commitment": {
+                    "inputHash": "f1620c3e929e13d3f0f55ce24e7579919e01b356e79b4212a622b4fc2e7b0acb10d0d",
+                    "outputHash": "f1620ff7f5beaf16900218a3ac4aae82cdccf764816986c7c739c716cf7dc03112a2c",
+                    "subQueriesHash": "f1620ca4510738395af1429224dd785675309c344b2b549632e20275c69b15ed1d210",
+                },
+                "proof": {
+                    "type": "Ed25519Signature2020",
+                    "verificationMethod": "did:key:z6Mko2nqhQ9wYSTS5Giab2j1aHzGnxHimqwmFeEVY8aNsVnN",
+                    "proofValue": "uZbm7fFcWc4l6iyvaKe_txdKntL3h3kvsGHOaKIbPV6c42PH1VnSmpYHMopv4TU67syzgoEdcS26AvpkSQb9dBQ",
+                }
+            })
+        );
+
+        // Successful validation
+        let mut request = response;
+        request.as_object_mut().unwrap().remove("output");
+        println!("{request:#?}");
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&request)
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap();
+
+        let response = res.json::<serde_json::Value>().await.unwrap();
+        pretty_assertions::assert_eq!(response, json!({"ok": true}));
+
+        // Invalid request: input hash
+        let mut invalid_request = request.clone();
+        invalid_request["commitment"]["inputHash"] =
+            "f1620c3e929e13d3f0f55ce24e7579919e01b356e79b4212a622b4fc2e7b0acb10d0e".into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "InvalidRequest::InputHash",
+                    "message": "The commitment is invalid and cannot be disputed: \
+                                commitment.inputHash doesn't match the hash of input object",
+                }
+            }),
+        );
+
+        // Invalid request: subQueries hash
+        let mut invalid_request = request.clone();
+        invalid_request["commitment"]["subQueriesHash"] =
+            "f1620ca4510738395af1429224dd785675309c344b2b549632e20275c69b15ed1d211".into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "InvalidRequest::SubQueriesHash",
+                    "message": "The commitment is invalid and cannot be disputed: \
+                                commitment.subQueriesHash doesn't match the hash of subQueries object",
+                }
+            }),
+        );
+
+        // Invalid request: bad signature
+        let mut invalid_request = request.clone();
+        invalid_request["proof"]["proofValue"] =
+            "uZbm7fFcWc4l6iyvaKe_txdKntL3h3kvsGHOaKIbPV6c42PH1VnSmpYHMopv4TU68syzgoEdcS26AvpkSQb9dBQ".into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "InvalidRequest::BadSignature",
+                    "message": "The commitment is invalid and cannot be disputed: \
+                                Verification equation was not satisfied",
+                }
+            }),
+        );
+
+        // Output mismatch
+        // (dataset stays the same but we fake the output hash and the signature)
+        let mut invalid_request = request.clone();
+        invalid_request["commitment"]["outputHash"] =
+            "f1620ff7f5beaf16900218a3ac4aae82cdccf764816986c7c739c716cf7dc03112a2d".into();
+        let c = canonical_json::to_string(&invalid_request["commitment"]).unwrap();
+        let sig: Signature = harness.private_key.sign(c.as_bytes()).into();
+        invalid_request["proof"]["proofValue"] = sig.to_string().into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "VerificationFailed::OutputMismatch",
+                    "actual_hash": "f1620ff7f5beaf16900218a3ac4aae82cdccf764816986c7c739c716cf7dc03112a2c",
+                    "expected_hash": "f1620ff7f5beaf16900218a3ac4aae82cdccf764816986c7c739c716cf7dc03112a2d",
+                    "message": "Query was reproduced but resulted in output hash different from expected. \
+                                This means that the output was either falsified, or the query \
+                                reproducibility was not guaranteed by the system.",
+                }
+            }),
+        );
+
+        // Cannot reproduce the query: Dataset is missing
+        // (dataset stays the same but we fake the output hash and the signature)
+        let mut invalid_request = request.clone();
+        invalid_request["input"]["datasets"][0]["id"] = DatasetID::new_seeded_ed25519(b"foo")
+            .as_did_str()
+            .to_string()
+            .into();
+        invalid_request["commitment"]["inputHash"] = Multihash::from_digest_sha3_256(
+            canonical_json::to_string(&invalid_request["input"])
+                .unwrap()
+                .as_bytes(),
+        )
+        .as_multibase()
+        .to_string()
+        .into();
+
+        let c = canonical_json::to_string(&invalid_request["commitment"]).unwrap();
+        let sig: Signature = harness.private_key.sign(c.as_bytes()).into();
+        invalid_request["proof"]["proofValue"] = sig.to_string().into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "VerificationFailed::DatasetNotFound",
+                    "dataset_id": "did:odf:fed01666f6fb3b7370000666f6fb3b737000060f6f60600000000895cddbcb7f7b8cc",
+                    "message": "Unable to reproduce the query as one of the input datasets cannot be found. The \
+                                owner of dataset either deleted it or made private or this node requires \
+                                additional configuration in order to locate it.",
+                }
+            }),
+        );
+
+        // Cannot reproduce the query: Block is missing
+        // (dataset stays the same but we fake the output hash and the signature)
+        let mut invalid_request = request.clone();
+        invalid_request["input"]["datasets"][0]["blockHash"] =
+            Multihash::from_digest_sha3_256(b"foo")
+                .as_multibase()
+                .to_string()
+                .into();
+        invalid_request["commitment"]["inputHash"] = Multihash::from_digest_sha3_256(
+            canonical_json::to_string(&invalid_request["input"])
+                .unwrap()
+                .as_bytes(),
+        )
+        .as_multibase()
+        .to_string()
+        .into();
+
+        let c = canonical_json::to_string(&invalid_request["commitment"]).unwrap();
+        let sig: Signature = harness.private_key.sign(c.as_bytes()).into();
+        invalid_request["proof"]["proofValue"] = sig.to_string().into();
+
+        let res = cl
+            .post(&format!("{}verify", harness.root_url))
+            .json(&invalid_request)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(res.status(), http::StatusCode::BAD_REQUEST);
+        pretty_assertions::assert_eq!(
+            res.json::<serde_json::Value>().await.unwrap(),
+            json!({
+                "ok": false,
+                "error": {
+                    "kind": "VerificationFailed::DatasetBlockNotFound",
+                    "block_hash": "f162076d3bc41c9f588f7fcd0d5bf4718f8f84b1c41b20882703100b9eb9413807c01",
+                    "dataset_id": "did:odf:fed01df230b49615d175307d580c33d6fda61fc7b9aec91df0f5c1a5ebe3b8cbfee02",
+                    "message": "Unable to reproduce the query as one of the input datasets does not contain a \
+                                block with specified hash. Under normal circumstances a block can disappear \
+                                only when the owner of dataset performs history-altering operation such as \
+                                reset or hard compation. There is also a probability that block hash was \
+                                spoofed in the original request to falsify the resuts.",
+                }
+            }),
+        );
     };
 
     await_client_server_flow!(harness.server_harness.api_server_run(), client);